requirements.txt
Django==5.0.6
psycopg2==2.9.9
pandas==2.2.2
djangorestframework==3.15.1
openpyxl==3.1.2
django-cors-headers==4.3.1